@@ -3,29 +3,29 @@
 	"private": true,
 	"dependencies": {
 		"@cospired/i18n-iso-languages": "^4.2.0",
-		"@sentry/react": "^7.110.0",
-		"antd": "5.16.1",
-		"dayjs": "^1.11.10",
+		"@sentry/react": "^7.112.2",
+		"antd": "5.16.5",
+		"dayjs": "^1.11.11",
 		"diff2html": "^3.4.48",
-		"dompurify": "^3.1.0",
+		"dompurify": "^3.1.2",
 		"filesize": "^10.1.1",
-		"i18n-iso-countries": "^7.11.0",
-		"i18next": "^23.11.1",
+		"i18n-iso-countries": "^7.11.1",
+		"i18next": "^23.11.3",
 		"lodash": "^4.17.21",
 		"minireset.css": "^0.0.7",
 		"path-browserify": "^1.0.1",
 		"pretty-bytes": "^6.1.1",
-		"react": "18.2.0",
+		"react": "18.3.1",
 		"react-aria-menubutton": "^7.0.3",
 		"react-audio-player": "^0.17.0",
 		"react-beautiful-dnd": "^13.1.1",
-		"react-dom": "18.2.0",
-		"react-i18next": "^14.1.0",
+		"react-dom": "18.3.1",
+		"react-i18next": "^14.1.1",
 		"react-image-crop": "^11.0.5",
-		"react-router-dom": "6.22.3",
+		"react-router-dom": "6.23.0",
 		"react-text-transition": "^3.1.0",
 		"react-toastify": "^10.0.5",
-		"react-virtuoso": "^4.7.8",
+		"react-virtuoso": "^4.7.10",
 		"resize-observer-polyfill": "^1.5.1",
 		"sass": "^1.75.0",
 		"slug": "^9.0.0",
@@ -36,12 +36,13 @@
 	"devDependencies": {
 		"@fortawesome/fontawesome-free": "^6.5.2",
 		"@types/lodash": "^4.17.0",
-		"@types/react": "18.2.78",
+		"@types/react": "18.3.1",
 		"@types/react-aria-menubutton": "^6.2.14",
 		"@types/react-beautiful-dnd": "^13.1.8",
-		"@types/react-dom": "18.2.25",
+		"@types/react-dom": "18.3.0",
 		"@typescript-eslint/eslint-plugin": "^7.0.2",
 		"@typescript-eslint/parser": "^7.0.2",
+		"@vitejs/plugin-legacy": "^5.4.0",
 		"@vitejs/plugin-react": "^4.2.1",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
@@ -52,8 +53,9 @@
 		"i18next-scanner": "^4.4.0",
 		"prettier": "^3.2.5",
 		"source-map-explorer": "^2.5.3",
+		"terser": "^5.31.0",
 		"typescript": "5.4.5",
-		"vite": "^5.2.8",
+		"vite": "^5.2.10",
 		"vite-plugin-node-polyfills": "^0.21.0"
 	},
 	"scripts": {
