@@ -160,61 +158,57 @@
     "yesql": "^7.0.0"
   },
   "devDependencies": {
-    "@electron/notarize": "^2.3.0",
-    "@sentry/cli": "^2.31.0",
+    "@electron/notarize": "^2.3.2",
+    "@sentry/cli": "^2.31.2",
     "@types/bcryptjs": "^2.4.6",
     "@types/better-queue": "^3.8.6",
-    "@types/chai": "^4.3.14",
+    "@types/chai": "^4.3.16",
     "@types/compression": "^1.7.5",
     "@types/cors": "^2.8.17",
     "@types/damerau-levenshtein": "^1.0.2",
     "@types/discord-rpc": "4.0.8",
-    "@types/dompurify": "^3.0.5",
     "@types/fs-extra": "^11.0.4",
     "@types/i18next-fs-backend": "^1.1.5",
     "@types/ip": "^1.1.3",
     "@types/js-yaml": "^4.0.9",
     "@types/jsonwebtoken": "^9.0.6",
     "@types/langs": "^2.0.5",
-    "@types/lodash": "4.17.0",
+    "@types/lodash": "4.17.4",
     "@types/mocha": "^10.0.6",
     "@types/multer": "^1.4.11",
-    "@types/node": "20.12.7",
-    "@types/pg": "8.11.2",
+    "@types/node": "20.12.12",
+    "@types/pg": "8.11.6",
     "@types/pg-copy-streams": "^1.2.5",
     "@types/qrcode": "^1.5.5",
     "@types/randomstring": "^1.3.0",
-    "@types/readline-sync": "^1.4.8",
     "@types/semver": "^7.5.8",
     "@types/slug": "^5.0.8",
     "@types/tmi.js": "^1.8.6",
     "@types/uuid": "^9.0.8",
     "@types/which": "^3.0.3",
     "@types/yesql": "^4.1.4",
-    "@typescript-eslint/eslint-plugin": "^7.7.0",
-    "@typescript-eslint/parser": "^7.7.0",
-    "chai": "^5.1.0",
-    "electron": "^29.2.0",
+    "chai": "^5.1.1",
+    "electron": "^30.0.7",
     "electron-builder": "^24.13.3",
-    "esbuild": "^0.20.2",
-    "eslint": "^8.57.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^18.0.0",
+    "esbuild": "^0.21.3",
+    "eslint": "^9.3.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-security": "^3.0.0",
     "eslint-plugin-simple-import-sort": "^12.1.0",
+    "globals": "^15.3.0",
     "husky": "^9.0.11",
     "i18next-scanner": "^4.4.0",
-    "lint-staged": "^15.2.2",
+    "lint-staged": "^15.2.4",
     "mocha": "^10.4.0",
-    "postgrator-cli": "^7.0.0",
+    "postgrator-cli": "^8.0.0",
     "prettier": "3.2.5",
-    "rimraf": "^5.0.5",
+    "rimraf": "^5.0.7",
     "ts-node": "^10.9.2",
     "typescript": "5.4.5",
-    "vitest": "^1.5.0"
+    "typescript-eslint": "^7.10.0",
+    "vitest": "^1.6.0"
   },
   "optionalDependencies": {
     "x11": "^2.3.0"
