@@ -452,7 +452,12 @@ class KaraDownload extends Component<unknown, KaraDownloadState> {
 							total: this.state.karasCount,
 							showQuickJumper: true,
 						}}
-						childrenColumnName="childrenColumnName"
+						scroll={{
+							x: true,
+						}}
+						expandable={{
+							showExpandColumn: false,
+						}}
 					/>
 				</Layout.Content>
 				{this.state.preview ? (
