@@ -50,7 +50,12 @@ class Viewcounts extends Component<unknown, ViewcountsState> {
 						dataSource={this.state.karas}
 						columns={this.columns}
 						rowKey="kid"
-						childrenColumnName="childrenColumnName"
+						scroll={{
+							x: true,
+						}}
+						expandable={{
+							showExpandColumn: false,
+						}}
 					/>
 				</Layout.Content>
 			</>
