@@ -118,7 +118,17 @@ function TagsDuplicate() {
 				description={i18next.t('HEADERS.TAG_DUPLICATES.DESCRIPTION')}
 			/>
 			<Layout.Content>
-				<Table dataSource={tags} columns={columns} rowKey="tid" />
+				<Table
+					dataSource={tags}
+					columns={columns}
+					rowKey="tid"
+					scroll={{
+						x: true,
+					}}
+					expandable={{
+						showExpandColumn: false,
+					}}
+				/>
 				<Modal
 					title={i18next.t('TAGS.TAG_DELETED_CONFIRM')}
 					open={deleteModal}
