@@ -16,6 +16,8 @@ function TagsList() {
 	const [searchParams, setSearchParams] = useSearchParams();
 
 	const [filter, setFilter] = useState(localStorage.getItem('tagFilter') || '');
+	const [currentPage, setCurrentPage] = useState(parseInt(localStorage.getItem('tagsPage')) || 1);
+	const [currentPageSize, setCurrentPageSize] = useState(parseInt(localStorage.getItem('tagsPageSize')) || 10);
 	const [tags, setTags] = useState<DBTag[]>([]);
 	const [tag, setTag] = useState<DBTag>();
 	const [deleteModal, setDeleteModal] = useState(false);
@@ -67,6 +69,13 @@ function TagsList() {
 	useEffect(() => {
 		refresh();
 	}, []);
+
+	const handleTableChange = pagination => {
+		setCurrentPage(pagination.current);
+		setCurrentPageSize(pagination.pageSize);
+		localStorage.setItem('tagsPage', pagination.current);
+		localStorage.setItem('tagsPageSize', pagination.pageSize);
+	};
 
 	const columns = [
 		{
@@ -159,7 +168,25 @@ function TagsList() {
 						})}
 					</Select>
 				</div>
-				<Table dataSource={tags} columns={columns} rowKey="tid" />
+				<Table
+					onChange={handleTableChange}
+					dataSource={tags}
+					columns={columns}
+					rowKey="tid"
+					scroll={{
+						x: true,
+					}}
+					expandable={{
+						showExpandColumn: false,
+					}}
+					pagination={{
+						position: ['topRight', 'bottomRight'],
+						current: currentPage || 1,
+						defaultPageSize: currentPageSize,
+						pageSize: currentPageSize,
+						showQuickJumper: true,
+					}}
+				/>
 				<Modal
 					title={i18next.t('TAGS.TAG_DELETED_CONFIRM')}
 					open={deleteModal}
