@@ -18,8 +18,8 @@ export function getDescriptionInLocale(settings: SettingsStoreData, description:
 		return description[user.main_series_lang]
 			? description[user.main_series_lang]
 			: description[user.fallback_series_lang]
-			  ? description[user.fallback_series_lang]
-			  : description.eng;
+				? description[user.fallback_series_lang]
+				: description.eng;
 	} else {
 		return description[getLanguageIn3B(langSupport)] ? description[getLanguageIn3B(langSupport)] : description.eng;
 	}
@@ -39,10 +39,10 @@ export function getTagInLanguage(
 		resulti18n = i18n[mainLanguage]
 			? i18n[mainLanguage]
 			: i18n[fallbackLanguage]
-			  ? i18n[fallbackLanguage]
-			  : i18n.eng
-			    ? i18n.eng
-			    : tag.name;
+				? i18n[fallbackLanguage]
+				: i18n.eng
+					? i18n.eng
+					: tag.name;
 	} else {
 		resulti18n = tag.name;
 	}
@@ -50,10 +50,10 @@ export function getTagInLanguage(
 		resultDescription = desc[mainLanguage]
 			? desc[mainLanguage]
 			: desc[fallbackLanguage]
-			  ? desc[fallbackLanguage]
-			  : desc.eng
-			    ? desc.eng
-			    : null;
+				? desc[fallbackLanguage]
+				: desc.eng
+					? desc.eng
+					: null;
 	}
 	return {
 		i18n: resulti18n,
@@ -94,8 +94,8 @@ export function getTitleInLocale(
 		return titles[user.main_series_lang]
 			? titles[user.main_series_lang]
 			: titles[user.fallback_series_lang]
-			  ? titles[user.fallback_series_lang]
-			  : titles[default_language];
+				? titles[user.fallback_series_lang]
+				: titles[default_language];
 	} else {
 		return titles[getLanguageIn3B(langSupport)] ? titles[getLanguageIn3B(langSupport)] : titles[default_language];
 	}
@@ -121,10 +121,10 @@ export function getSeriesSingersFull(settings: SettingsStoreData, data: DBKara, 
 	return data?.series?.length > 0
 		? data.series.map(e => getTagInLocale(settings, e, i18nParam).i18n).join(', ')
 		: data?.singergroups?.length > 0
-		  ? data.singergroups.map(e => getTagInLocale(settings, e, i18nParam).i18n).join(', ')
-		  : data?.singers?.length > 0
-		    ? data.singers.map(e => getTagInLocale(settings, e, i18nParam).i18n).join(', ')
-		    : ''; // wtf?;
+			? data.singergroups.map(e => getTagInLocale(settings, e, i18nParam).i18n).join(', ')
+			: data?.singers?.length > 0
+				? data.singers.map(e => getTagInLocale(settings, e, i18nParam).i18n).join(', ')
+				: ''; // wtf?;
 }
 
 /**
@@ -144,27 +144,27 @@ export function buildKaraTitle(
 		data.langs = [isMulti];
 	}
 	const serieText =
-		data.from_display_type && data[data.from_display_type]
+		data.from_display_type && data[data.from_display_type] && data[data.from_display_type].length > 0
 			? data[data.from_display_type]
 					.slice(0, 3)
 					.map(e => getTagInLocale(settings, e, i18nParam).i18n)
 					.join(', ') + (data[data.from_display_type].length > 3 ? '...' : '')
 			: data?.series?.length > 0
-			  ? data.series
+				? data.series
 						.slice(0, 3)
 						.map(e => getTagInLocale(settings, e, i18nParam).i18n)
 						.join(', ') + (data.series.length > 3 ? '...' : '')
-			  : data?.singergroups?.length > 0
-			    ? data.singergroups
+				: data?.singergroups?.length > 0
+					? data.singergroups
 							.slice(0, 3)
 							.map(e => getTagInLocale(settings, e, i18nParam).i18n)
 							.join(', ') + (data.singergroups.length > 3 ? '...' : '')
-			    : data?.singers?.length > 0
-			      ? data.singers
+					: data?.singers?.length > 0
+						? data.singers
 								.slice(0, 3)
 								.map(e => getTagInLocale(settings, e, i18nParam).i18n)
 								.join(', ') + (data.singers.length > 3 ? '...' : '')
-			      : ''; // wtf?
+						: ''; // wtf?
 	const langsText = data?.langs
 		?.map(e => e.name)
 		.join(', ')
@@ -343,7 +343,7 @@ export function computeTagsElements(kara: DBKara, scope: Scope, versions = true,
 					<div key={isMulti.tid} className="tag black">
 						{getInlineTag(isMulti, tagTypes.LANGS.type, scope, i18nParam)}
 					</div>
-			  )
+				)
 			: karaTags.push(
 					...sortAndHideTags(kara.langs, scope).map(tag => {
 						return (
@@ -352,7 +352,7 @@ export function computeTagsElements(kara: DBKara, scope: Scope, versions = true,
 							</div>
 						);
 					})
-			  );
+				);
 	}
 	if (kara.songtypes) {
 		karaTags.push(
@@ -424,7 +424,7 @@ export function computeTagsElements(kara: DBKara, scope: Scope, versions = true,
 														</span>
 													),
 													x,
-											  ],
+												],
 									null
 								)}
 						</span>
