@@ -1,12 +1,13 @@
-import { defineConfig } from 'vite';
+import legacy from '@vitejs/plugin-legacy';
 import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite';
 import { nodePolyfills } from 'vite-plugin-node-polyfills';
 
 export default defineConfig({
 	build: {
 		sourcemap: true,
 	},
-	plugins: [nodePolyfills(), react()],
+	plugins: [nodePolyfills(), react(), legacy()],
 	server: {
 		port: 3000,
 		proxy: {
