@@ -35,7 +35,10 @@ export async function getInbox(repoName: string, token: string): Promise<Inbox[]
 				authorization: token,
 			},
 		});
-		const availableKaras = await getKarasMicro(res.data.flatMap(d => [d.kid, d.edited_kid]));
+		const availableKaras = await getKarasMicro(
+			res.data.flatMap(d => [d.kid, d.edited_kid]),
+			true
+		);
 		return res.data.map(resdata => ({
 			...resdata,
 			available_locally: availableKaras.some(kara => kara.kid === resdata.kid || kara.kid === resdata.edited_kid),
@@ -72,7 +75,7 @@ export async function downloadKaraFromInbox(inid: string, repoName: string, toke
 					service,
 					obj: err,
 				});
-				throw new ErrorKM('INBOX_VIEW_ERROR');
+				throw new ErrorKM('INBOX_VIEW_ERROR', 500, false);
 			}
 		}
 		// If song has a parent in the inbox and we don't have it yet, download it first.
