@@ -9,33 +9,35 @@ import { removeParentInKaras } from '../dao/karafile.js';
 import { selectPlaylistContentsMicro } from '../dao/playlist.js';
 import { saveSetting } from '../lib/dao/database.js';
 import { refreshKarasDelete } from '../lib/dao/kara.js';
-import { formatKaraV4, getDataFromKaraFile, writeKara } from '../lib/dao/karafile.js';
+import { extractMediaTechInfos, formatKaraV4, getDataFromKaraFile, writeKara } from '../lib/dao/karafile.js';
 import { refreshTags } from '../lib/dao/tag.js';
 import { writeTagFile } from '../lib/dao/tagfile.js';
+import { APIMessage } from '../lib/services/frontend.js';
 import { refreshKarasAfterDBChange, updateTags } from '../lib/services/karaManagement.js';
 import { DBKara, DBKaraTag } from '../lib/types/database/kara.js';
 import { DBTag } from '../lib/types/database/tag.js';
 import { KaraFileV4, KaraTag } from '../lib/types/kara.js';
 import { TagTypeNum } from '../lib/types/tag.js';
 import { ASSFileSetMediaFile } from '../lib/utils/ass.js';
-import { resolvedPathRepos } from '../lib/utils/config.js';
+import { resolvedPath, resolvedPathRepos } from '../lib/utils/config.js';
 import { getTagTypeName } from '../lib/utils/constants.js';
 import { ErrorKM } from '../lib/utils/error.js';
 import { fileExists, resolveFileInDirs } from '../lib/utils/files.js';
 import logger, { profile } from '../lib/utils/logger.js';
+import { encodeMediaToRepoDefault } from '../lib/utils/mediaInfoValidation.js';
 import { createImagePreviews } from '../lib/utils/previews.js';
 import Task from '../lib/utils/taskManager.js';
+import { emitWS } from '../lib/utils/ws.js';
 import { adminToken } from '../utils/constants.js';
 import sentry from '../utils/sentry.js';
 import { getState } from '../utils/state.js';
 import { checkMediaAndDownload } from './download.js';
 import { getKara, getKaras } from './kara.js';
 import { editKara } from './karaCreation.js';
-import { getRepo, getRepos } from './repo.js';
+import { getRepo, getRepoManifest, getRepos } from './repo.js';
 import { updateAllSmartPlaylists } from './smartPlaylist.js';
 import { getTag } from './tag.js';
-import { APIMessage } from '../lib/services/frontend.js';
-import { emitWS } from '../lib/utils/ws.js';
+import i18next from 'i18next';
 
 const service = 'KaraManager';
 
@@ -68,6 +70,8 @@ export async function removeKara(
 		const parents: Family[] = [];
 		const karas = await selectAllKaras({
 			q: `k:${kids.join(',')}`,
+			ignoreCollections: true,
+			blacklist: false,
 		});
 		if (karas.length === 0) throw new ErrorKM('UNKNOWN_SONG', 404, false);
 		for (const kara of karas) {
@@ -106,6 +110,8 @@ export async function removeKara(
 					parent: kara.kid,
 					children: await selectAllKaras({
 						q: `k:${kara.children.join(',')}`,
+						ignoreCollections: true,
+						blacklist: false,
 					}),
 				});
 			}
@@ -339,6 +345,54 @@ export async function deleteMediaFile(file: string, repo: string) {
 		logger.error(`Unable to delete media file ${file} from repository ${repo} : ${err}`, { service });
 		sentry.error(err);
 		throw new ErrorKM('MEDIA_DELETE_ERROR');
+	}
+}
+
+export async function encodeMediaFileToRepoDefaults(
+	kid?: string,
+	tempFileName?: string,
+	repo?: string,
+	encodeOptions?: { trim?: boolean },
+	task = new Task({
+		value: 0,
+	})
+) {
+	task.update({
+		text: 'CALCULATING_MEDIA_ENCODING_PARAMETERS',
+		value: 0,
+	});
+
+	try {
+		// It's okay to not have a kid (when we get a new karaoke)
+		const kara = kid ? await getKara(kid, adminToken) : null;
+		const mediaFilePaths =
+			(tempFileName && [resolve(resolvedPath('Temp'), basename(tempFileName))]) ||
+			(await resolveFileInDirs(kara.mediafile, resolvedPathRepos('Medias', kara.repository)));
+		const mediaFileExists = mediaFilePaths.length > 0 && (await fileExists(mediaFilePaths[0]));
+		if (!mediaFileExists) throw new ErrorKM('Mediafile not found');
+		const currentMediaInfo =
+			mediaFilePaths.length > 0 &&
+			(await fileExists(mediaFilePaths[0])) &&
+			(await extractMediaTechInfos(mediaFilePaths[0]));
+		task.update({
+			text: 'ENCODING_MEDIA',
+			subtext: kid ? kara.titles[kara.titles_default_language] : i18next.t('UNKNOWN'),
+			total: currentMediaInfo.duration,
+		});
+		const repoManifest = getRepoManifest(repo ?? kara.repository);
+		const encodedFileInfo = await encodeMediaToRepoDefault(mediaFilePaths[0], currentMediaInfo, repoManifest, {
+			trim: encodeOptions?.trim,
+			outputFolder: resolvedPath('Temp'),
+			onProgress: progress => {
+				task.update({
+					value: progress.timeSeconds,
+				});
+			},
+		});
+		const newMediaInfo = await extractMediaTechInfos(encodedFileInfo.newMediaFilePath, undefined, true);
+		return newMediaInfo;
+	} finally {
+		task.end();
 	}
 }
 
