@@ -83,6 +83,8 @@ import { getUser, getUsers, updateSongsLeft } from './user.js';
 
 const service = 'Playlist';
 
+let freeOrphanedSongsIntervalID;
+
 /** Test if basic playlists exist */
 export async function testPlaylists() {
 	profile('testPlaylists');
@@ -423,6 +425,7 @@ function currentHook(plaid: string, name: string) {
 	setState({ currentPlaid: plaid, introPlayed: false, introSponsorPlayed: false });
 	emitWS('currentPlaylistUpdated', plaid);
 	resetAllAcceptedPLCs();
+	writeStreamFiles('next_song_name_and_requester');
 	writeStreamFiles('current_kara_count');
 	writeStreamFiles('time_remaining_in_current_playlist');
 	downloadMediasInPlaylist(plaid);
@@ -828,6 +831,7 @@ export async function addKaraToPlaylist(params: AddKaraParams) {
 		const plc = await getPLCInfo(PLCsInserted[0].plcid, true, requester);
 		if (params.plaid === state.currentPlaid) {
 			checkMediaAndDownload(karas);
+			writeStreamFiles('next_song_name_and_requester');
 			writeStreamFiles('current_kara_count');
 			writeStreamFiles('time_remaining_in_current_playlist');
 			if (conf.Karaoke.Autoplay && (state.player?.mediaType === 'stop' || state.randomPlaying)) {
@@ -1046,6 +1050,7 @@ export async function removeKaraFromPlaylist(
 			}
 		}
 		await Promise.all([
+			writeStreamFiles('next_song_name_and_requester'),
 			writeStreamFiles('current_kara_count'),
 			writeStreamFiles('time_remaining_in_current_playlist'),
 			writeStreamFiles('public_kara_count'),
@@ -1221,6 +1226,7 @@ export async function editPLC(plc_ids: number[], params: PLCEditParams, refresh 
 			if (currentSong && currentSong.pos <= params.pos && plc.plaid === getState().currentPlaid) {
 				setState({ player: { currentSong: await getCurrentSong() } });
 			}
+			writeStreamFiles('next_song_name_and_requester');
 			writeStreamFiles('time_remaining_in_current_playlist');
 			songsLeftToUpdate.add({
 				username: plc.username,
@@ -1671,13 +1677,17 @@ async function freeOrphanedSongs() {
 /** Initialize playlist tasks */
 export async function initPlaylistSystem() {
 	profile('initPL');
-	setInterval(freeOrphanedSongs, 60 * 1000);
+	freeOrphanedSongsIntervalID = setInterval(freeOrphanedSongs, 60 * 1000);
 	const pls = await selectPlaylists(false);
 	pls.forEach(pl => reorderPlaylist(pl.plaid));
 	await testPlaylists();
 	updateAllSmartPlaylists();
 	logger.debug('Playlists initialized', { service });
 	profile('initPL');
+}
+
+export function stopPlaylistSystem() {
+	if (freeOrphanedSongsIntervalID) clearInterval(freeOrphanedSongsIntervalID);
 }
 
 /** Update all user quotas affected by a PLC getting freed/played */
